# public/openapi.yaml
openapi: 3.0.3
info:
  title: Jotty API
  description: |
    Programmatic access to your checklists and notes with full CRUD operations.

    ## Authentication
    All endpoints require authentication via API key in the `x-api-key` header.

    Get your API key from: Profile → Settings → API Key → Generate
  version: "1.0.0"
  contact:
    name: Jotty
    url: https://github.com/fccview/jotty
security:
  - ApiKeyAuth: []
paths:
...
  /tasks:
    get:
      summary: Get All Tasks
      description: Retrieves all task checklists for the authenticated user. Supports filtering and search.
      tags:
        - Tasks
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter tasks by category
        - name: status
          in: query
          schema:
            type: string
          description: Filter tasks that contain items with this status
        - name: q
          in: query
          schema:
            type: string
          description: Search tasks by title or item text
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                        title:
                          type: string
                          example: "Sprint Tasks"
                        category:
                          type: string
                          example: "Work"
                        statuses:
                          type: array
                          items:
                            $ref: '#/components/schemas/KanbanStatus'
                        items:
                          type: array
                          items:
                            $ref: '#/components/schemas/ChecklistItem'
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time

    post:
      summary: Create Task
      description: Creates a new task checklist for the authenticated user
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: "New Sprint"
                category:
                  type: string
                  default: "Uncategorized"
                  example: "Work"
                statuses:
                  type: array
                  items:
                    $ref: '#/components/schemas/KanbanStatus'
                  description: Custom Kanban statuses (defaults to todo/in_progress/completed)
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      title:
                        type: string
                      category:
                        type: string
                      statuses:
                        type: array
                        items:
                          $ref: '#/components/schemas/KanbanStatus'
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ChecklistItem'
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        '400':
          description: Invalid request
...
  /tasks/{taskId}/statuses:
    get:
      summary: Get Task Statuses
      description: Retrieves all Kanban column statuses for a task
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the task
      responses:
        '200':
          description: List of statuses
          content:
            application/json:
              schema:
                type: object
                properties:
                  statuses:
                    type: array
                    items:
                      $ref: '#/components/schemas/KanbanStatus'
        '404':
          description: Task not found
        '400':
          description: Not a task checklist

    post:
      summary: Create Status
      description: Adds a new Kanban column status to a task
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - label
              properties:
                id:
                  type: string
                  example: "review"
                label:
                  type: string
                  example: "In Review"
                color:
                  type: string
                  example: "#3b82f6"
                order:
                  type: integer
                  example: 2
      responses:
        '200':
          description: Status created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/KanbanStatus'
        '400':
          description: Invalid request or status ID already exists
        '404':
          description: Task not found
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key obtained from Profile → Settings → API Key
  schemas:
...
    KanbanStatus:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the status
          example: "todo"
        label:
          type: string
          description: Display name for the status
          example: "To Do"
        color:
          type: string
          description: Optional color code for the status column
          example: "#3b82f6"
        order:
          type: integer
          description: Display order of the status column
          example: 0
    ChecklistItem:
      type: object
      properties:
        id:
          type: string
          description: Internal item ID
          example: "listId-1234567890"
        index:
          type: integer
          description: Position in the parent's items array
          example: 0
        text:
          type: string
          example: "Task 1"
        completed:
          type: boolean
          example: false
        status:
          type: string
          enum: [todo, in_progress, paused, completed]
          description: Only present for task-type checklists
        time:
          oneOf:
            - type: integer
              example: 0
            - type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  startTime:
                    type: string
                    format: date-time
                  endTime:
                    type: string
                    format: date-time
                  duration:
                    type: integer
          description: Only present for task-type checklists
        children:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
          description: Nested sub-items
...